EPMRPP-79077 || fix failed to start suit error.
Also user points that he has the same error when he tried to run default
test with generates by playwright itself. The main problem was when
onTestEnd hook runs, we decrease testsLength property only in rootTestsObject,
and in object, with contains information only about test with currently ends.
Previously it decrease the testsLength in all testsObjects.